+++ spec/factories/user.rb
@@ -13,6 +13,12 @@ FactoryBot.define do
       # nothing on purpose
     end
 
+    trait :password_reseted do
+      after(:create) do |user|
+        user.generate_reset_password_token!
+      end
+    end
+
     trait :not_accepted_tos do
       terms_of_service nil
     end