+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### API changes
 
 - Provide `POST /api/users/password_resets` endpoint
+- Provide `POST /api/users/passwords` endpoint
 
 ## Aries 2018-08-27